issue #112 rename : MakeMemoService -> DetailPlanService
