+++ client/.idea/misc.xml
@@ -24,6 +24,7 @@
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/calendar_day_layout.xml" value="0.3854166666666667" />
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/dialog_join_plan.xml" value="0.28072916666666664" />
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/dialog_remove_confirm.xml" value="0.2713541666666667" />
+        <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/dialog_search_flight.xml" value="0.215625" />
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/fragment_add_place.xml" value="0.21458333333333332" />
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/fragment_airline_search.xml" value="0.26197916666666665" />
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/fragment_base.xml" value="0.38645833333333335" />
