+++ client/app/src/main/java/com/example/witt/data/source/remote/memo/MakeMemoDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.example.witt.data.source.remote.memo
 
-import com.example.witt.data.api.PlanMemoService
+import com.example.witt.data.api.DetailPlanService
 import com.example.witt.data.model.plan.detail_plan.request.MemoRequest
 import com.example.witt.data.model.plan.memo.MakeMemoResponse
 import com.example.witt.di.IoDispatcher
@@ -9,7 +9,7 @@ import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class MakeMemoDataSourceImpl @Inject constructor(
-    private val service: PlanMemoService,
+    private val service: DetailPlanService,
     @IoDispatcher private val coroutineDispatcher: CoroutineDispatcher
 ) : MakeMemoDataSource{
